moved @types/moo to dev depenedncies
