+++ packages/lang-server/src/token.ts
@@ -17,10 +17,7 @@ import os from 'os'
 import { parser } from '@salto-io/workspace'
 import { EditorPosition } from './context'
 
-export type Token = {
-  value: string
-  type: string
-}
+export type Token = Pick<parser.Token, 'value' | 'type'>
 
 export const getToken = (fileContent: string, position: EditorPosition):
   Token | undefined => {
@@ -32,21 +29,14 @@ export const getToken = (fileContent: string, position: EditorPosition):
   // and cause us to not support multiline tokens
   const line = lines[position.line]
 
-  const lexer = new parser.PeekableLexer(line)
-  try {
-    while (true) {
-      const token = lexer.next()
-      const col = token.col - 1
-      if (col <= position.col && position.col < col + token.value.length) {
-        return { value: token.value, type: token.type }
-      }
-      if (col > position.col) {
-        return undefined
-      }
+  for (const token of parser.tokenizeContent(line)) {
+    const col = token.col - 1
+    if (col <= position.col && position.col < col + token.value.length) {
+      return { value: token.value, type: token.type }
     }
-  // eslint-disable-next-line no-empty
-  } catch (e) {
-    // Lexer throws an error when there is nothing left to parse
+    if (col > position.col) {
+      return undefined
+    }
   }
   return undefined
 }
