+++ packages/workspace/package.json
@@ -40,7 +40,6 @@
     "jest-each": "^26.5.0",
     "lodash": "^4.17.19",
     "moo": "^0.5.1",
-    "@types/moo": "^0.5.0",
     "nearley": "2.19.2",
     "wu": "^2.1.0"
   },
@@ -56,6 +55,7 @@
     "@types/wu": "^2.1.40",
     "@typescript-eslint/eslint-plugin": "2.31.0",
     "@typescript-eslint/parser": "2.31.0",
+    "@types/moo": "^0.5.0",
     "eslint": "^6.2.2",
     "eslint-config-airbnb": "18.0.1",
     "eslint-plugin-header": "^3.0.0",
