+++ packages/workspace/src/parser/index.ts
@@ -13,10 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import PeekableLexer, { LexerToken } from './internal/native/lexer'
-
-export { parse, SourceRange, parseElemID } from './parse'
+export { parse, SourceRange, parseElemID, tokenizeContent, Token } from './parse'
 export { ParseResult, ParseError } from './types'
 export { dumpElements, dumpElemID, dumpValues } from './dump'
 export { SourceMap } from './source_map'
-export { PeekableLexer, LexerToken }
